Refactor and remove direct dbus dependency
