+++ wpinet/build.gradle
@@ -172,11 +172,7 @@ ext {
                     }
                 }
             }
-            it.cppCompiler.args.addAll([
-                '-I/usr/include/dbus-1.0',
-                '-I/usr/lib/x86_64-linux-gnu/dbus-1.0/include'
-            ])
-            it.linker.args << '-ldbus-1'
+            it.linker.args << '-ldl'
         }
     }
 }
