+++ wpinet/wpinet-config.cmake.in
@@ -1,7 +1,6 @@
 include(CMakeFindDependencyMacro)
 @FILENAME_DEP_REPLACE@
 @LIBUV_SYSTEM_REPLACE@
-@DBUS_DEP_REPLACE@
 @WPIUTIL_DEP_REPLACE@
 
 @FILENAME_DEP_REPLACE@
